Modified ShapeParent to reproduce the same reuse/de-duplication of THREE ShapeGeometries that the NodeView class did when all of the core node drawing operations were contained in it.  Also, added a number of comments and notes.
